requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.8
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.212
langchainplus-sdk==0.0.17
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pydantic==1.10.9
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
